# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: NumCheck API
  description: |
    Event-driven, multi-tenant system for bulk validation of WhatsApp & Telegram numbers.
    
    ## Authentication
    
    The API supports two authentication methods:
    1. **JWT Token**: Use `Authorization: Bearer <token>` header
    2. **API Key**: Use `X-Api-Key: <api-key>` header
    
    ## Rate Limiting
    
    - Quick check: 60 requests/minute per tenant
    - Bulk operations: 200 items/second total processing
    - Rate limit headers are included in responses
    
    ## Error Handling
    
    The API uses standard HTTP status codes and returns error details in JSON format:
    ```json
    {
      "error": "Error message",
      "code": "ERROR_CODE",
      "details": {}
    }
    ```
  version: 1.0.0
  contact:
    name: NumCheck Support
    email: support@numcheck.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.numcheck.com
    description: Production
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      summary: User login
      description: Authenticate user with email and password
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: demo@example.com
                password:
                  type: string
                  example: demo123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/me:
    get:
      summary: Get current user
      description: Get information about the authenticated user
      tags: [Authentication]
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/admin/tenants:
    get:
      summary: List tenants
      description: Get a list of all tenants (admin only)
      tags: [Admin - Tenants]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  tenants:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      summary: Create tenant
      description: Create a new tenant (admin only)
      tags: [Admin - Tenants]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tenant'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/admin/users:
    get:
      summary: List users
      description: Get a list of users (admin only)
      tags: [Admin - Users]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: tenantId
          in: query
          schema:
            type: string
          description: Filter by tenant ID
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
  /api/admin/tg-accounts:
    get:
      summary: List Telegram accounts
      description: Get a list of Telegram accounts for validation
      tags: [Admin - Telegram]
      responses:
        '200':
          description: List of Telegram accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TelegramAccount'

    post:
      summary: Add Telegram account
      description: Add a new Telegram account for validation
      tags: [Admin - Telegram]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTelegramAccountRequest'
      responses:
        '201':
          description: Telegram account added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramAccount'
...
  /api/admin/jobs:
    get:
      summary: List all jobs
      description: Get a list of all validation jobs (admin only)
      tags: [Admin - Jobs]
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed, canceled]
        - name: tenantId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Job'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [ADMIN, TENANT_ADMIN, USER]
        tenantId:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        plan:
          type: string
        rateLimit:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Job:
      type: object
      properties:
        id:
          type: string
        tenantId:
          type: string
        type:
          type: string
          enum: [QUICK, BULK]
        status:
          type: string
          enum: [PENDING, RUNNING, COMPLETED, FAILED, CANCELED]
        total:
          type: integer
        processed:
          type: integer
        success:
          type: integer
        failed:
          type: integer
        createdAt:
          type: string
          format: date-time
    TelegramAccount:
      type: object
      properties:
        id:
          type: string
        phoneLabel:
          type: string
        isActive:
          type: boolean
        dailyLimit:
          type: integer
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        errorCount:
          type: integer
    CreateTenantRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        plan:
          type: string
          default: "basic"
        rateLimit:
          type: integer
          minimum: 1
          maximum: 10000
          default: 60
    CreateTelegramAccountRequest:
      type: object
      required: [phoneLabel, apiId, apiHash, sessionString]
      properties:
        phoneLabel:
          type: string
        apiId:
          type: string
        apiHash:
          type: string
        sessionString:
          type: string
        proxyUrl:
          type: string
          format: uri
          nullable: true
        dailyLimit:
          type: integer
          minimum: 1
          maximum: 10000
          default: 1000
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
        details:
          type: object
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Validation failed"
            code: "VALIDATION_ERROR"
            details:
              field: "numbers"
              message: "Array must contain at least 1 element"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authentication required"
            code: "UNAUTHORIZED"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Insufficient permissions"
            code: "FORBIDDEN"
...